function_apps/func-end-function/requirements.txt
azure-functions
pytest==7.4.2
black==23.11.0
azure.core==1.29.5
azure.data.tables==12.4.0
python-dotenv==1.0.0
azure.storage.blob
openpyxl
pandas==2.1.1
